added function updateTable which changes all tables in tree according to passed param newTables
Pedro, you might have a problem with this implementation :)
but it seems that removeNaturalJoin table also assumes that we are only working with natural joins so this isn't limiting us in any way i think.
but let me know what you think.
also, getTablesFromCols is always retuning an empty array.