fixed issue with tables not getting removed when removing column
fixed getTablesFromCols and a bunch of other stuff.

the tables get removed correctly now but there is still a problem with the actual columns in the UI not getting removed.
it seems like it's an issue with the final few lines in removeSelectColumn, especially:
//Convert to * optimized column list
    let newColumns = colsToStar(columns, newTables)

    //Add columns to the tree
    newTree.value.selectItems.value = newColumns.map(mapColString)

I will look into it  next.